refactor(tangle-dapp): Better names for `AmountInput` and `AddressInput` components
