+++ apps/tangle-dapp/components/AmountInput/useInputAmount.tsx
